requirements.txt
Flask==2.0.1
pymongo==3.12.0
python_bcrypt==0.3.2
gunicorn